pyproject.toml
[project]
name = "deep-research-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.7.4",
    "autogen-ext[langchain,openai]>=0.7.4",
    "ddgs>=0.3.0",
    "langchain-community>=0.3.29",
    "python-dotenv>=1.1.1",
]

[project.scripts]
deep-research-bot = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
